- draw forest + test, fix init of burning trees

+++ src/main/java/fr/gillesphan/cirilgroup/config/AppConfiguration.java
@@ -17,13 +17,18 @@ public class AppConfiguration {
 
     private int width;
     private int height;
+    /**
+     * To delete ?
+     * 
+     * @deprecated not used
+     */
     private List<FirePosition> startFirePositions;
     private double contagionRate;
 
     public AppConfiguration() {
     }
 
-    public AppConfiguration(int width, int height, List<FirePosition> startFirePositions, int contagionRate) {
+    public AppConfiguration(int width, int height, List<FirePosition> startFirePositions, double contagionRate) {
         this.width = width;
         this.height = height;
         this.startFirePositions = startFirePositions;
